Rename base class file
