+++ dgm_utils/model.py
